+++ test/sceneloader.cpp
@@ -63,8 +63,8 @@ void openGfx(eve::window::Window *&win, eve::graphics::Graphics *&gfx, int w = 3
     REQUIRE(win != nullptr);
     REQUIRE(gfx != nullptr);
     eve::window::WindowSettings s;
-    s.width = w;
-    s.height = h;
+    s.width  = static_cast<uint16_t>(w);
+    s.height = static_cast<uint16_t>(h);
     s.centered = true;
     REQUIRE(win->setWindowSettings(s));
 }
